requirements.txt
flask
flask-cors
flask-bcrypt
flask-mail
mysql-connector-python
itsdangerous
gunicorn